pyproject.toml
[tool.poetry]
name = "gitlab-tools"
version = "0.1.7"
description = ""
authors = ["Yangming Huang <yangming@inceptioglobal.ai>"]
packages = [
    { include = "gitlab_tools" },
]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.22.0"
click = "^7.0"
pykebab = "^0.2.6"

[tool.poetry.scripts]
lab = 'gitlab_tools.cli:main'

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
